refactored all views to seperate files
