+++ Greetings/MessageModel.swift
@@ -1,0 +1,14 @@
+//
+//  MessageModel.swift
+//  Greetings
+//
+//  Created by Guy Steinberg on 2024-04-14.
+//
+
+import SwiftUI
+
+struct DataItemModel: Identifiable {
+    let id = UUID()
+    let text: LocalizedStringKey
+    let color: Color
+}
